refactor: simplify app structure and remove unused
