+++ web-application/front-app/src/App.css
@@ -11,81 +11,10 @@ body {
   font-weight: 300;
 }
 
-.navBar {
-  position: fixed;
-  top: 0;
-  z-index: 1;
-  display: flex;
-  width: 100%;
-  height: 60px;
-  background: rgba(0, 0, 0, 0.9);
-}
-
-.navBar ul {
-  display: flex;
-  list-style: none;
-  width: 100%;
-  justify-content: center;
-}
-
-.navBar li {
-  padding: 1rem;
-  margin: 0 1px;
-  /* background-color: #333; */
-}
-
-.navBar ul li a {
-  text-decoration: none;
-  padding-bottom: 20px;
-  text-transform: uppercase;
-  color: #f4f4f4;
-}
-
-.navBar ul li a:hover {
-  color: #e69500;
-  border-bottom: 3px solid orange;
-}
-
 .App {
   text-align: center;
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
-}
-
-section {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-  width: 100%;
-  height: 100vh;
-}
-
-button {
-  margin: 8px;
-  padding: 2px;
-}
-
-input {
-  margin: 3px;
-}
-
-@media only screen and (max-width: 915px) {
-  .navBar ul li a {
-    font-size: 12px;
-  }
-  .navBar ul li a:hover {
-    color: #e69500;
-    border-bottom: none;
-  }
-}
-
-#githubLink {
-  transition: all 0.2s ease-in-out;
-}
-#githubLink:hover {
-  transform: scale(1.1);
 }
