--- web-application/front-app/src/components/Home.js
@@ -1,11 +1,0 @@
-import React from "react";
-import logo from "../images/tableau.png";
-
-export default function Home() {
-  return (
-    <div>
-      <img src={logo} alt="tableauImage" />
-      <h1>React Playground for Tableau JavaScript API</h1>
-    </div>
-  );
-}
